Prep: Move mk_tempfile to testfile::file
