+++ src/rust/engine/testutil/Cargo.toml
@@ -13,4 +13,5 @@ grpc_util = { path = "../grpc_util" }
 hashing = { path = "../hashing" }
 prost = { workspace = true }
 protos = { path = "../protos" }
+tempfile = { workspace = true }
 tokio = { workspace = true }
