+++ src/rust/engine/testutil/src/file.rs
@@ -4,6 +4,8 @@ use std::io::Read;
 use std::os::unix::fs::PermissionsExt;
 use std::path::Path;
 
+use tokio::io::{AsyncSeekExt, AsyncWriteExt};
+
 pub fn list_dir(path: &Path) -> Vec<String> {
   let mut v: Vec<_> = std::fs::read_dir(path)
     .unwrap_or_else(|err| panic!("Listing dir {path:?}: {err:?}"))
@@ -31,4 +33,19 @@ pub fn is_executable(path: &Path) -> bool {
   std::fs::metadata(path)
     .map(|meta| meta.permissions().mode() & 0o100 == 0o100)
     .unwrap_or(false)
+}
+
+pub async fn mk_tempfile(contents: Option<&[u8]>) -> tokio::fs::File {
+  let file = tokio::task::spawn_blocking(tempfile::tempfile)
+    .await
+    .unwrap()
+    .unwrap();
+  let mut file = tokio::fs::File::from_std(file);
+
+  if let Some(contents) = contents {
+    file.write_all(contents).await.unwrap();
+    file.rewind().await.unwrap();
+  }
+
+  file
 }
